requirements.txt
streamlit
pdfplumber
langchain-groq
langchain-core
matplotlib
kiwipiepy
pandas
pandas
bcrypt
gspread
google-auth
